refactor(akrites): move scatter/legend VM interfaces to shared
LFXV2-2366
